# openapi/switchboard.openapi.json
{
  "components": {
    "schemas": {
...
      "TaskCreateCommandV1": {
        "additionalProperties": false,
        "description": "Transport-neutral input for creating one task.",
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Assignee"
          },
          "deliverable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Deliverable"
          },
          "depends_on": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Depends On",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Description"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Duration Days"
          },
          "effort_days": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Effort Days"
          },
          "entry_criteria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Entry Criteria"
          },
          "exit_criteria": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Exit Criteria"
          },
          "finish_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Finish Date"
          },
          "is_blocking": {
            "default": false,
            "title": "Is Blocking",
            "type": "boolean"
          },
          "owner_org": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Owner Org"
          },
          "owner_person_or_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Owner Person Or Role"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Phase"
          },
          "risk_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Risk Level"
          },
          "schema": {
            "default": "switchboard.task.create_command.v1",
            "title": "Schema",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Start Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workstream_id": {
            "title": "Workstream Id",
            "type": "string"
          },
          "workstream_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Workstream Name"
          }
        },
        "required": [
          "workstream_id",
          "title"
        ],
        "title": "CreateTaskCommand",
        "type": "object"
      },
...
      "TaskMoveCommandV1": {
        "additionalProperties": false,
        "description": "Transport-neutral input for moving one task between project boards.",
        "properties": {
          "dependency_policy": {
            "default": "fail",
            "title": "Dependency Policy",
            "type": "string"
          },
          "new_task_id": {
            "default": "",
            "title": "New Task Id",
            "type": "string"
          },
          "project_from": {
            "title": "Project From",
            "type": "string"
          },
          "project_to": {
            "title": "Project To",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "schema": {
            "default": "switchboard.task.move_command.v1",
            "title": "Schema",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "project_from",
          "project_to"
        ],
        "title": "MoveTaskCommand",
        "type": "object"
      },
      "TaskUpdateCommandV1": {
        "additionalProperties": false,
        "description": "Transport-neutral input for a sparse task update.",
        "properties": {
          "depends_on": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Depends On"
          },
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          },
          "schema": {
            "default": "switchboard.task.update_command.v1",
            "title": "Schema",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id"
        ],
        "title": "UpdateTaskCommand",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Contract-first REST surface for Switchboard. Request bodies for the paths below are generated from ``src/switchboard/contracts`` Pydantic models shared with MCP application commands.",
    "title": "Switchboard REST",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/tasks": {
      "post": {
        "operationId": "createTask",
        "parameters": [
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional retry key. Same key + same body replays the original response; same key + different body returns 409 idem_key_conflict.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateCommandV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "error_code": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "idem_key": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Idempotency-Key conflict (idem_key_conflict)"
          }
        },
        "summary": "Create a task",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "getTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          }
        },
        "summary": "Get one task",
        "tags": [
          "contracts"
        ]
      },
      "patch": {
        "operationId": "updateTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional retry key. Same key + same body replays the original response; same key + different body returns 409 idem_key_conflict.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdateCommandV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "error_code": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "idem_key": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Idempotency-Key conflict (idem_key_conflict)"
          }
        },
        "summary": "Update a task",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/tasks/{task_id}/move": {
      "post": {
        "operationId": "moveTask",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional retry key. Same key + same body replays the original response; same key + different body returns 409 idem_key_conflict.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskMoveCommandV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Invalid request payload"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "error_code": {
                      "const": "idem_key_conflict",
                      "type": "string"
                    },
                    "idem_key": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "operation": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Idempotency-Key conflict (idem_key_conflict)"
          }
        },
        "summary": "Move a task between project boards",
        "tags": [
          "contracts"
        ]
      }
    },
...
  }
}